Clarify user goal prompt in Rule 1 Apply section
BEFORE:
'In one sentence, write the user goal. Start with 'The user can…' or 'The user should be able to…''

AFTER:
'In one sentence, write the user goal. Start with 'The user can…' or 'The user should be able to…' (where 'user' refers to the person doing the action)'

This makes it clear that 'user' refers to the person performing the action, not just using the word 'user' everywhere.